+++ android/src/main/kotlin/it/mainella/phone_state/receiver/PhoneStateReceiver.kt
@@ -7,8 +7,21 @@ import android.telephony.TelephonyManager
 import it.mainella.phone_state.utils.PhoneStateStatus
 
 open class PhoneStateReceiver : BroadcastReceiver() {
-    var status: PhoneStateStatus = PhoneStateStatus.NOTHING;
-    var phoneNumber: String? = null;
+    var status: PhoneStateStatus = PhoneStateStatus.NOTHING
+    var phoneNumber: String? = null
+
+    fun instance(context: Context) {
+        val telephonyManager = context.getSystemService(Context.TELEPHONY_SERVICE) as TelephonyManager
+
+        status = when (telephonyManager.callState) {
+            TelephonyManager.CALL_STATE_RINGING -> PhoneStateStatus.CALL_INCOMING
+            TelephonyManager.CALL_STATE_OFFHOOK -> PhoneStateStatus.CALL_STARTED
+            TelephonyManager.CALL_STATE_IDLE -> PhoneStateStatus.CALL_ENDED
+            else -> PhoneStateStatus.NOTHING
+        }
+        phoneNumber = null
+    }
+
     override fun onReceive(context: Context?, intent: Intent?) {
         try {
             status = when (intent?.getStringExtra(TelephonyManager.EXTRA_STATE)) {