+++ example/android/app/src/main/kotlin/it/mainella/example/MainActivity.kt
@@ -1,6 +1,0 @@
-package it.mainella.example
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}