+++ example/android/app/src/main/kotlin/it/mainella/phone_state_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package it.mainella.phone_state_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity()