+++ lib/phone_state.dart
@@ -1,4 +1,4 @@
-library phone_state;
+library;
 
 export 'src/phone_state.dart' show PhoneState;
 export 'src/utils/phone_state_status.dart' show PhoneStateStatus;