+++ lib/src/phone_state.dart
@@ -1,3 +1,4 @@
+import 'package:collection/collection.dart';
 import 'package:flutter/services.dart';
 import 'package:phone_state/src/utils/constants.dart';
 import 'package:phone_state/src/utils/phone_state_status.dart';
@@ -17,21 +18,17 @@ class PhoneState {
   /// This method allows you to create a [PhoneState] object with the status [PhoneStateStatus.NOTHING]
   ///
   /// Use for initializing your [PhoneState] object
-  factory PhoneState.nothing() =>
-      PhoneState._(status: PhoneStateStatus.NOTHING);
+  factory PhoneState.nothing() => PhoneState._(status: PhoneStateStatus.NOTHING);
 
-  static const EventChannel _eventChannel =
-      EventChannel(Constants.EVENT_CHANNEL);
+  static const EventChannel _eventChannel = EventChannel(Constants.EVENT_CHANNEL);
 
-  /// This method allows you to have a stream of the system phone state change
-  static Stream<PhoneState> get stream {
-    return _eventChannel
-        .receiveBroadcastStream()
-        .distinct()
-        .map((dynamic event) => PhoneState._(
-              status: PhoneStateStatus.values.firstWhere(
-                  (element) => element.name == event['status'] as String),
-              number: event['phoneNumber'],
-            ));
-  }
+  /// This variable allows you to have a stream of the system phone state change
+  static final Stream<PhoneState> stream =
+      _eventChannel.receiveBroadcastStream().distinct().map((dynamic event) => PhoneState._(
+            status: PhoneStateStatus.values.firstWhereOrNull(
+                  (element) => element.name == event['status'] as String,
+                ) ??
+                PhoneStateStatus.NOTHING,
+            number: event['phoneNumber'],
+          ));
 }